requirements.txt
numpy
scipy
torchio
torch
logging
joblib
dipy
matplotlib
scienceplots
cvxpy
matlabengine